Derive finances header state reactively from transactions
