+++ app/src/main/java/org/ole/planet/myplanet/model/FinanceHeaderState.kt
@@ -1,0 +1,8 @@
+package org.ole.planet.myplanet.model
+
+data class FinanceHeaderState(
+    val debit: Int = 0,
+    val credit: Int = 0,
+    val total: Int = 0,
+    val isCautionVisible: Boolean = false
+)
